refactor: extract videoGen BYOV runtime mgmt + history I/O out of local.js (#1833)
